Introduce \App\Core\BaseController, as base Class for both Classic and Themed Controllers
