+++ app/Core/ClassicController.php
@@ -10,21 +10,16 @@
 namespace App\Core;
 
 use Smvc\Core\View;
-use Smvc\Core\Controller;
+use App\Core\BaseController;
 
 /**
  * Simple themed controller showing the typical usage of the Flight Control method.
  */
-class ClassicController extends Controller
+class ClassicController extends BaseController
 {
     protected $layout = 'default';
 
-    // Store the Controller's variables.
-    protected $data = array();
-
-    protected $autoRender = true;
-    protected $useLayout  = false;
-
+    
     /**
      * Call the parent construct
      */
@@ -53,15 +48,13 @@ class ClassicController extends Controller
         if(($result === true) || is_null($result)) {
             $data =& $this->data();
 
-            $content = View::render($this->method(), $data, false, true);
-
             if($this->useLayout) {
+                $content = View::render($this->method(), $data, false, true);
+
                 View::renderLayout($this->layout(), $content, $data);
             }
             else {
-                View::sendHeaders();
-
-                echo $content;
+                View::render($this->method(), $data, false);
             }
 
             // The current Page was rendered there; stop the Flight.
@@ -70,63 +63,6 @@ class ClassicController extends Controller
 
         // Leave to the parent's method the Flight decisions.
         return parent::afterFlight($result);
-    }
-
-    public function autoRender($value = null)
-    {
-        if(is_null($value)) {
-            return $this->autoRender;
-        }
-
-        $this->autoRender = $value;
-    }
-
-    public function useLayout($value = null)
-    {
-        if(is_null($value)) {
-            return $this->useLayout;
-        }
-
-        $this->useLayout = $value;
-    }
-
-    public function data($name = null)
-    {
-        if(is_null($name)) {
-            return $this->data;
-        }
-        else if(isset($this->data[$name])) {
-            return $this->data[$name];
-        }
-
-        return null;
-    }
-
-    public function set($name, $value = null)
-    {
-        if (is_array($name)) {
-            if (is_array($value)) {
-                $data = array_combine($name, $value);
-            }
-            else {
-                $data = $name;
-            }
-        }
-        else {
-            $data = array($name => $value);
-        }
-
-        $this->data = $data + $this->data;
-    }
-
-    public function title($title)
-    {
-        $data = array('title' => $title);
-
-        $this->data = $data + $this->data;
-
-        // Auto-activate the Rendering on Layout.
-        $this->useLayout = true;
     }
 
 }
